--- Struts2/T1/strutsT1/src/struts.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE struts PUBLIC
-    "-//Apache Software Foundation//DTD Struts Configuration 2.5//EN"
-    "http://struts.apache.org/dtds/struts-2.5.dtd">
-<struts>
-	<package name = "default" 	extends = "struts-default">
-		<action name = "helloworld" class = "com.sacura.t1.action.HelloAction">
-			<result name = "success">/jsp/hello.jsp</result>
-		</action>
-	</package>
-</struts>
