+++ Struts2/T1/任增超/strutsT1/.settings/org.eclipse.wst.jsdt.ui.superType.name
@@ -1,0 +1,1 @@
+Window
