+++ Struts2/T1/任增超/strutsT1/WebContent/jsp/hello.jsp
@@ -1,0 +1,14 @@
+<%@ page language="java" contentType="text/html; charset=utf-8"
+    pageEncoding="utf-8"%>
+<%@taglib uri="/struts-tags" prefix="s" %>
+<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
+<html>
+<head>
+<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
+<title>Insert title here</title>
+</head>
+<body>
+	<p style="color: red"><s:property value="context"/></p>
+	<p style="color: blue">${context }</p>
+</body>
+</html>
